+++ docs/en/changes/changes.md
@@ -7,6 +7,7 @@
 * BanyanDB: Support `hot/warm/cold` stages configuration.
 * Fix query continues profiling policies error when the policy is already in the cache.
 * Support `hot/warm/cold` stages TTL query in the status API.
+* PromQL Service: traffic query support `limit` and regex match.
 
 #### UI